Improved README readablity
Added a section about what the theme actually does.
Updated links to guides to point to what will be the urls in the new_guides branch.
Removed section about creating a dummy test app, this belongs in the guide.
Added a section about creating your own extensions, with links to the guide.